+++ tests/cogs/test_owner_cog.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from functools import partial
+from inspect import cleandoc
 from typing import Any
 from unittest.mock import AsyncMock, MagicMock
 
@@ -8,18 +9,18 @@ import pytest
 from discord.ext import commands
 from pytest_mock import MockerFixture
 
-from spellbot.cogs import BanCog
+from spellbot.cogs import OwnerCog
 from spellbot.database import DatabaseSession
 from spellbot.models import User
 from tests.mixins import ContextMixin
 
 
 @pytest.mark.asyncio
-class TestCogBan(ContextMixin):
+class TestCogOwner(ContextMixin):
     async def run(
         self,
         cog: commands.Cog,
-        func: commands.Command[BanCog, ..., None],
+        func: commands.Command[OwnerCog, ..., None],
         *args: Any,
         **kwargs: Any,
     ) -> None:
@@ -29,7 +30,7 @@ class TestCogBan(ContextMixin):
     async def test_ban_and_unban(self) -> None:
         target_user = MagicMock()
         target_user.id = 1002
-        cog = BanCog(self.bot)
+        cog = OwnerCog(self.bot)
 
         await self.run(cog, cog.ban, self.context, str(target_user.id))
 
@@ -49,12 +50,12 @@ class TestCogBan(ContextMixin):
         assert not users[0].banned
 
     async def test_ban_without_target(self) -> None:
-        cog = BanCog(self.bot)
+        cog = OwnerCog(self.bot)
         await self.run(cog, cog.ban, self.context, None)
         self.context.author.send.assert_called_once_with("No target user.")
 
     async def test_ban_with_invalid_target(self) -> None:
-        cog = BanCog(self.bot)
+        cog = OwnerCog(self.bot)
         await self.run(cog, cog.ban, self.context, "abc")
         self.context.author.send.assert_called_once_with("Invalid user id.")
 
@@ -65,9 +66,9 @@ class TestCogBan(ContextMixin):
     ) -> None:
         target_user = MagicMock()
         target_user.id = 1002
-        cog = BanCog(self.bot)
+        cog = OwnerCog(self.bot)
 
-        mocker.patch("spellbot.cogs.ban_cog.set_banned", AsyncMock(side_effect=RuntimeError()))
+        mocker.patch("spellbot.cogs.owner_cog.set_banned", AsyncMock(side_effect=RuntimeError()))
 
         with pytest.raises(RuntimeError):
             await self.run(cog, cog.ban, self.context, str(target_user.id))
@@ -78,3 +79,22 @@ class TestCogBan(ContextMixin):
         with pytest.raises(RuntimeError):
             await self.run(cog, cog.unban, self.context, str(target_user.id))
         assert "rolling back database session due to unhandled exception" in caplog.text
+
+    async def test_stats(self) -> None:
+        cog = OwnerCog(self.bot)
+
+        await self.run(cog, cog.stats, self.context)
+
+        self.context.author.send.assert_called_once_with(
+            cleandoc(
+                """
+                    ```
+                    status:   online
+                    activity: None
+                    ready:    False
+                    guilds:   0
+                    users:    0
+                    ```
+                """
+            )
+        )