+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds owner command: !stats.
+
 ## [v8.3.9](https://github.com/lexicalunit/spellbot/releases/tag/v8.3.9) - 2022-11-25
 
 ### Fixed